+++ server/src/main/java/com/krillsson/sysapi/core/domain/system/SystemInfo.kt
@@ -21,7 +21,6 @@
 package com.krillsson.sysapi.core.domain.system
 
 import com.krillsson.sysapi.core.domain.cpu.CpuInfo
-import com.krillsson.sysapi.core.domain.drives.Drive
 import com.krillsson.sysapi.core.domain.filesystem.FileSystem
 import com.krillsson.sysapi.core.domain.gpu.Gpu
 import com.krillsson.sysapi.core.domain.memory.MemoryInfo
@@ -35,7 +34,6 @@ class SystemInfo(
     val cpuInfo: CpuInfo,
     val motherboard: Motherboard,
     val memory: MemoryInfo,
-    val drives: List<Drive>,
     val fileSystems: List<FileSystem>,
     val networkInterfaces: List<NetworkInterface>,
     val gpus: List<Gpu>